cmd/cl-adm: Move core logic to a separate package
This commit moves the core logic of the cl-adm command to a pkg/bootstrap.
pkg/bootstrap implements a memory-based bootstrap, instead of the previous file-based
approach which was implemented in cl-adm.
The cl-adm command still gives a file-based API, which is wrapped around the memory-based
API that is implemented in pkg/bootstrap.

Signed-off-by: Or Ozeri <oro@il.ibm.com>