Add version info package
- populates source control variables from runtime/debug.BuildInfo
- adds a global version.txt file (should be updated on every formal release)
- Short() to allow setting cobra.Command.Version string

The placement of version.txt in the pkg/versioninfo directory is unfortunate.
Unfortunately go does not allow ".." in embed paths. A possible workaround is to
use go:generate comment to copy the file from the root to this location (and adding
the local version to .gitignore...)
An alternative would be to use a go:generate comment with "git describe" to find
the latest tag reachable from the current revision..

Signed-off-by: Etai Lev Ran <elevran@gmail.com>