controlplane: Store JWK keys in k8s secret
Today, the JWK is initialized by the controlplane when loading,
and stored in-memory.
To support multiple controlpanes with the same JWK key,
we change the controlplane authz manager to ready the JWK key
from a secret, instead of generating it.manager
The secret is created by the (soon to be leader elected) control manager.

Signed-off-by: Or Ozeri <oro@il.ibm.com>